requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
accelerate==0.23.0
h5py==3.7.0
matplotlib==3.5.2
pandas==1.4.4
einops==0.7.0
timm==0.9.7
tensorboard==2.14.1
yaml==0.2.5
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.9.1
torch-optimizer==0.3.0

